feat: add timeline for pills today
