+++ src/components/organisms/WeeklyPill/index.jsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState } from 'react'
-import { useSelector } from 'react-redux'
+import React from 'react'
+import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import Cross from '../../atoms/Cross'
 import Title from '../../molecules/Title'
@@ -14,42 +14,7 @@ const dayOfWeek = {
   일: 0,
 }
 
-const WeeklyPill = () => {
-  const pills = useSelector((state) => state.pills.pills)
-  const loading = useSelector((state) => state.pills.loading)
-  const weeklyPills = useSelector((state) => state.pills.weeklyPills) // today가 아니라 week(그것도 day/요일)임
-  const [dayPills, setDayPills] = useState([...weeklyPills])
-
-  console.log(dayPills)
-  console.log(weeklyPills)
-  useEffect(() => {
-    const filteredPills = pills.filter((pill) => pill.freqDay !== 0)
-    filteredPills.forEach((pill) => {
-      if (pill.freqDay === '1' || pill.freqDay === 1) {
-        setDayPills((prev) => prev.map((el) => [...el, pill]))
-      } else {
-        // 생성 날짜로부터 오늘날짜 기준 월~일에 어느 때 먹어야 하는지 계산
-        const today = new Date()
-        const pillDayArr = pill.created?.toDate().toLocaleDateString('ko-KO').split('. ')
-
-        const [year, month, day] = [today.getFullYear(), today.getMonth() + 1, today.getDate()]
-        const dayOffset = today.getDay() === 0 ? 6 : today.getDay() - 1 // 일요일 7로 바꿈
-
-        const thenDate = new Date(...pillDayArr)
-        const monDate = new Date(year, month, day - dayOffset) // 그 주의 월요일만 구함
-
-        const btMs = monDate.getTime() - thenDate.getTime()
-        const btDay = btMs / (1000 * 60 * 60 * 24)
-
-        setDayPills((prev) =>
-          prev.map((el, index) =>
-            (btDay + index) % Number(pill.freqDay) === 0 ? [...el, pill] : el,
-          ),
-        )
-      }
-    })
-  }, []) // 바로 업데이트가 되지 않음
-
+const WeeklyPill = ({ pills }) => {
   return (
     <>
       <Title title="Weekly" infoText="" toggle={false} />
@@ -58,9 +23,9 @@ const WeeklyPill = () => {
           <Weekday key={day}>
             <h4>{day}</h4>
             <ThumbPill>
-              {dayPills.length > 0
-                ? `${dayPills[index][0]?.name ?? '없음'} ${
-                    dayPills[index].length > 1 ? `외 ${dayPills[index].length - 1} 종` : ''
+              {pills.length > 0
+                ? `${pills[index][0]?.name ?? '없음'} ${
+                    pills[index].length > 1 ? `외 ${pills[index].length - 1} 종` : ''
                   } `
                 : '없음'}
             </ThumbPill>
@@ -107,5 +72,9 @@ const ThumbPill = styled.div`
   font-size: 12px;
   text-align: center;
 `
+
+WeeklyPill.propTypes = {
+  pills: PropTypes.array.isRequired,
+}
 
 export default WeeklyPill
